Changes by harrison

+++ server/pre_project/src/main/java/com/codestates/backend/pre_project/post/comment/service/CommentService.java
@@ -2,6 +2,7 @@ package com.codestates.backend.pre_project.post.comment.service;
 
 import com.codestates.backend.pre_project.exception.BusinessLogicException;
 import com.codestates.backend.pre_project.exception.ExceptionCode;
+import com.codestates.backend.pre_project.member.service.MemberService;
 import com.codestates.backend.pre_project.post.comment.entity.Comment;
 import com.codestates.backend.pre_project.post.comment.repository.CommentRepository;
 import com.codestates.backend.pre_project.utils.CustomBeanUtils;
@@ -9,6 +10,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import com.codestates.backend.pre_project.member.entity.Member;
 
 import java.util.Optional;
 
@@ -18,10 +20,12 @@ public class CommentService {
     private final CommentRepository commentRepository;
 
     private final CustomBeanUtils<Comment> beanUtils;
+    private final MemberService memberService;
 
-    public CommentService(CommentRepository commentRepository, CustomBeanUtils<Comment> beanUtils) {
+    public CommentService(CommentRepository commentRepository, CustomBeanUtils<Comment> beanUtils, MemberService memberService) {
         this.commentRepository = commentRepository;
         this.beanUtils = beanUtils;
+        this.memberService = memberService;
     }
 
     public Comment createComment(Comment comment) {
@@ -31,6 +35,9 @@ public class CommentService {
 
     public Comment updateComment(Comment comment) {
         Comment findComment = findVerifiedComment(comment.getCommentId());
+        Member commentMember = memberService.findVerifiedMember(findComment.getMember().getMemberId());
+        if (memberService.getCurrentMember().getMemberId() != commentMember.getMemberId())
+            throw new BusinessLogicException(ExceptionCode.EDIT_NOT_ALLOWED);
 
         Comment updateComment = beanUtils.copyNonNullProperties(comment, findComment);
 
@@ -46,6 +53,10 @@ public class CommentService {
 
     public void deleteComment(long commentId) {
         Comment findComment = findVerifiedComment(commentId);
+        Member commentMember = memberService.findVerifiedMember(findComment.getMember().getMemberId());
+        if (memberService.getCurrentMember().getMemberId() != commentMember.getMemberId())
+            throw new BusinessLogicException(ExceptionCode.EDIT_NOT_ALLOWED);
+//        작성자에게만 수정 권한을 주는 코드. 멤버서비스에 코드 구현 필요
 
         commentRepository.delete(findComment);
     }
